Fix a minor build issue on OS X
This commit removes a minor build issue that pops up on OS X. Most of the build
issues are around the use of `libc::ucred` in ancillary.rs which is not exposed
by OS X's libc. This commit does nothing for that, instead resolving a
simple-to-address problem with a function being await that doesn't need to be,
hidden on Linux owing to feature flags.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>