Add DogStatsD as a Regression Detector target
This commit adds DogStatsD experiments to the regression detector setup. This
has not been wired into CI as of this commit, however I intend to do that in
follow-up work in this PR. Unfortunately the dogstatsd container as configured
exits with 0 immediately, meaning no experimental data is collected. Repro:

```
> RUST_LOG=debug ~/PATH/TO/smp local-run --target-image datadog/dogstatsd:7.52.1 --experiment-dir `pwd`/test/smp/regression/dogstatsd --case uds_dogstatsd_to_api --total-samples 60
```

I don't know enough about this target to debug why it exits immediately and
silently.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>