[APR-205] dogstatsd: add support for a configurable tag interceptor for filtering/augmenting metadata via tags (#132)
This PR introduces a new trait, `TagMetadataInterceptor`, which serves two purposes:
- evaluate individual tags to determine whether they should be passed through, dropped, or "intercepted"
- when a tag is intercepted, define how we use that tag to augment the metadata for the metric

Most of the code involved here is boilerplate to fit how we've structured decoding to avoid allocations when building `ContextRef<'a, T>`... so we've like `TagSplitter<'a>`/`TagIter<'a>`, we now have `TagFilterer<I, TMI>`/`TagFiltererIter<I, TMI>`.

We handle this in two phases:

- wrap our normal `TagSplitter<'a>` to get `TagFilterer<TagSplitter<'a>, TMI>`, where `TMI` is the interceptor implementation to use (this can be cloned, and serves purely as a way to filter out intercepted/dropped tags when doing context stuff)
- call a new function, `update_metadata_from_tags`, which handles updating the `MetricMetadata` we got from decoding with any tags that get identified as needing to be intercepted

For now, we've pushed this into `DogstatsdCodec` as a new generic parameter (`TMI`) with a default of `()`. All the relevant traits implementations are in place _such_ that existing code works exactly as it did before, and no tags are filtered out or do anything to modify the metric metadata. In a subsequent PR, we'll add a Agent-like interceptor implementation specifically for handling some of the tags we currently handle in the `origin_enrichment` transform, such as `dd.internal.entity_id`, `dd.internal.jxm_check_name`, and `dd.internal.card`.