Start to flatten out framing and codecs
The goal here is just to start reducing some complexity so that subsequent changes are easier to make and simpler to review. Where previously we had a `Deserializer` wrapping a `Framer` wrapping a `Codec`, we now just have a framer and a codec that are used side-by-side. By avoiding the nesting, we're making it easier to introduce a semi-parsed result struct from the codec without needing to thread reference lifetimes through multiple layers of machinery. Actually implementing that change to the codec's parsing function will be the next step, and from there we will be able to rework the rest of the logic (enrichment, differentiating between message types, forwarding, etc) to be a more straightforward function from that semi-parsed representation to a full event, with fewer intermediate steps and less mutation, as well as our ultimate goal, to easily pass in a string/context interner specific to that socket handler.

Signed-off-by: Luke Steensen <luke.steensen@gmail.com>