Add ability to register ADP as a remote agent to the Datadog Agent (#377)
# Context

We would like to be able to register ADP as a remote agent to the Core
Agent so that status and flares can be requested.

# Solution

This pr creates a new destination component that is intended for both
status and flare information.

# Notes

This pr registers ADP as a remote agent but a following pr is needed for
actually sending data over.

# Testing

Run ADP with an agent that has the remote agent registry enabled. 
Run the `status` command on the agent. You should see ADP listed in the
`Remote Agents` section.

---------

Signed-off-by: Raymond Zhao <raymond.zhao@datadoghq.com>