enhancement: add proxy support (#530)
## Summary
<!-- Please provide a brief summary about what this PR does.
This should help the reviewers give feedback faster and with higher
quality. -->
This pr adds the use of `ProxyConnector` from `hyper-proxy2` in order to
proxy requests sent by `HttpClient`.

`ProxyConnector` is wrapped around the existing `HttpsCapableConnector`

Note: This pr does not included feature of `no_proxy` yet, which is used
to bypass the proxy for certain hosts.

## Change Type
- [ ] Bug fix
- [x] New feature
- [ ] Non-functional (chore, refactoring, docs)
- [ ] Performance


## How did you test this PR?
<!-- Please how you tested these changes here -->
1. Set up
[squid](https://docs.datadoghq.com/agent/configuration/proxy/?tab=linux#squid)
to use as a forward proxy.
2. Set `DD_PROXY_HTTP` / `DD_PROXY_HTTPS` to point the server
3. Send metrics
4. Verify that the metrics are still being sent properly and check
`/var/log/squid/access.log` to see that the request is actually being
proxied.
## References

<!-- Please list any issues closed by this PR. -->

<!--
- Closes: <issue link>
-->

<!-- Any other issues or PRs relevant to this PR? Feel free to list them
here. -->

---------

Signed-off-by: Raymond Zhao <raymond.zhao@datadoghq.com>