enhancement: use zstd by default for metric payloads (#554)
## Summary

Add support for zstd compression of metric payloads and make it the
default behavior to match the Agent.

This required a change to how our chunked buffers accept writes, as we
were seeing some deadlock-like behavior with the previous implementation
that waited to collect the full capacity for each write. This PR changes
that to write as much as it can to the available buffers, even if that
is not the full amount, and allow the rest to be written on subsequent
calls.

## Change Type
- [ ] Bug fix
- [x] New feature
- [ ] Non-functional (chore, refactoring, docs)
- [ ] Performance


## How did you test this PR?
<!-- Please how you tested these changes here -->

## References

<!-- Please list any issues closed by this PR. -->
AGTMETRICS-56
<!--
- Closes: <issue link>
-->

<!-- Any other issues or PRs relevant to this PR? Feel free to list them
here. -->

---------

Signed-off-by: Luke Steensen <luke.steensen@gmail.com>