chore: chunk up the processing of event buffers in datadog events/service checks destination. (#555)
## Summary
<!-- Please provide a brief summary about what this PR does.
This should help the reviewers give feedback faster and with higher
quality. -->
This pr updates the datadog events/servicechecks destination to match
the behavior of the metrics destination to simply grab the next event
buffer instead of grabbing all available buffers.

## Change Type
- [ ] Bug fix
- [ ] New feature
- [x] Non-functional (chore, refactoring, docs)
- [ ] Performance

## How did you test this PR?
<!-- Please how you tested these changes here -->
Sent event and service check to datadog and made sure it still worked.
## References

<!-- Please list any issues closed by this PR. -->

<!--
- Closes: <issue link>
-->

<!-- Any other issues or PRs relevant to this PR? Feel free to list them
here. -->

Signed-off-by: Raymond Zhao <raymond.zhao@datadoghq.com> a2b5daa80b0e087c832bfda2ace020e772c62c0d