enhancement: Add abillity to send telemetry to the datadog agent (#582)
## Summary
<!-- Please provide a brief summary about what this PR does.
This should help the reviewers give feedback faster and with higher
quality. -->
This pr adds the ability for the `RemoteAgentImpl` to respond to the
`GetTelemetry` rpc call from the datadog-agent.

A request is made to the prometheus listen addr only if telemetry is
enabled.
## Change Type
- [ ] Bug fix
- [x] New feature
- [ ] Non-functional (chore, refactoring, docs)
- [ ] Performance

## How did you test this PR?
<!-- Please how you tested these changes here -->
1. Run the agent
2. Run ADP and wait for it to register as a remote agent
3. Go to the agent's telemetry endpoint localhost:5000/telemetry
endpoint and see that ADP's telemetry is there.
## References

<!-- Please list any issues closed by this PR. -->

<!--
- Closes: <issue link>
-->

<!-- Any other issues or PRs relevant to this PR? Feel free to list them
here. -->

This pr depends on https://github.com/DataDog/datadog-agent/pull/35319 0f4d1a0898efd3f2a57493069d4f781007e9bc3d