fix(ci): pass proper env var for ADP binary 'full name' value (#633)
## Summary

In CI, we use the `emit-build-metadata` Make target to generate the
"build metadata" that gets bundled into the ADP binary: build data,
short/full name of the binary, and so on. We were incorrectly passing
the "full name" to the relevant Docker build step (`APP_FULLNAME`
instead of `APP_FULL_NAME`), leading to the default value of `unknown`
being used.

This manifests as showing a name of `unknown` when ADP connects to the
Core Agent's "Remote Agent Registry", as well as a User Agent of
`unknown/0.1.0` when sending outbound HTTP requests, and so on.

## Change Type

- [x] Bug fix
- [ ] New feature
- [ ] Non-functional (chore, refactoring, docs)
- [ ] Performance

## How did you test this PR?

Will require waiting for CI to build an image that we can test locally.

## References

N/A