[AGENTRUN-227] chore(checks-agent): compare check-agent in Go against Rust version (#636)
This PR benchmarks the Go version of the checks agent
https://github.com/DataDog/datadog-agent/pull/31571 using SMP. We mirror
the image created in the PR to the SMP ECR. Since the PR has few
changes, we use the same commit as the baseline and comparison.

We updated the links generated to use the Go version as the comparison
for the SMP dashboard.