enhancement: add support for `histogram_copy_to_distribution` feature (#657)
## Summary
<!-- Please provide a brief summary about what this PR does.
This should help the reviewers give feedback faster and with higher
quality. -->
This pr adds support for the `histogram_copy_to_distribution_feature`.
The copy is done in the `aggregate` transform.
## Change Type
- [ ] Bug fix
- [x] New feature
- [ ] Non-functional (chore, refactoring, docs)
- [ ] Performance


## How did you test this PR?
<!-- Please how you tested these changes here -->
Set `histogram_copy_to_distribution` to `true` and set
`histogram_copy_to_distribution_prefix`. Emit a histogram metric and see
that the histogram statistics show up and as well as a new distribution
metrics on the Datadog UI

## References

<!-- Please list any issues closed by this PR. -->

<!--
- Closes: <issue link>
-->
- Closes: #646

<!-- Any other issues or PRs relevant to this PR? Feel free to list them
here. -->

---------

Signed-off-by: Raymond Zhao <raymond.zhao@datadoghq.com>