chore(datadog): refactor `RequestBuilder<O>` to be trait-driven to allow reuse between destinations (#658)
## Summary

This PR introduces an updated/refactored version of `RequestBuilder<O>` (from the Datadog Metrics destination) that is meant to be shared amongst all Datadog destinations. This refactored version allows for the encoding logic (among other things) to be pluggable using a new trait, `EndpointEncoder`.

The general idea is that each destination will implement the trait to define what the endpoint URI for payloads should be, how to encode each individual event that gets pushed into the request builder, the content type, size limits, and so on. `RequestBuilder<E, O>` then becomes responsible for the generic logic of managing compression, compression buffers, request splitting, and request building.

This is in preparation of cutting over the Datadog Events/Service Checks destination to use the request builder infrastructure, as we're planning to add support for multiple events per request, along with using a buffer pool for request building, and taking advantage of the existing work around this is preferable.

## Change Type

- [ ] Bug fix
- [ ] New feature
- [x] Non-functional (chore, refactoring, docs)
- [ ] Performance

## How did you test this PR?

Basic unit tests as this is simply duplicating the `RequestBuilder<O>` code for use in follow-up PRs.

## References

AGTMETRICS-184