chore(datadog): add support for writing delimited payloads to `RequestBuilder<E, O>` (#660)
## Summary

This PR adds support for delimited payloads to `RequestBuilder<E, O>`.

In the case of the Datadog Metrics destination, individual inputs are self-delimiting by virtue of being elements in a repeated message field in the Protocol Buffers payload. This means that while the request itself is a single message with N inputs bundled into it, and not N messages, we still don't have to do anything special to wrap those inputs or separate them: just write them individually, concatenated, and the payload is valid.

However, for other destinations, like Datadog Events/Service Checks, the payloads are JSON arrays. This means that the payload has a prefix and suffix (`[` and `]`) and each input needs to be separated by a comma. Naturally, you get this when encoding an array of values all at once, but we want to build our requests incrementally, input by input.

This PR adds support for this model by allowing `EndpointEncoder` to specify a prefix, suffix, and input separator (referred to collectively as "delimiters") that must be used when building a payload. `RequestBuilder<E, O>` uses these to generate a valid payload, whether we've written a single input or 10,000 inputs.

While these changes are simple, most of this PR revolves around ensuring that our size limiting logic and request splitting logic works correctly in delimited mode as well as non-delimited mode. We've done a decent amount of refactoring to try and share more of the encode/flush logic between the regular encode/flush functions, and their usage in the request splitting codepath.

## Change Type

- [ ] Bug fix
- [ ] New feature
- [x] Non-functional (chore, refactoring, docs)
- [ ] Performance

## How did you test this PR?

Unit tests and correctness test.

## References

AGTMETRICS-184 ad9cec336f4a2519a5f531d0c5285ace97288610