[AGENTRUN-287] feature(checks-agent): add  checks source and scheduler (#662)
This adds the checks source and scheduler.

The check's source reads events from the autodiscovery provider, and
based on the type of event, it would schedule or unschedule events into
the scheduler.

The check collector maintains a list of all checks, stored in a HashSet
to ensure that duplicated events, likely to occur from a reconnect, do
not schedule duplicate checks. c5d1d0ac831e9ac834c00213181954bb965d7700