track length of interner separate from offset into backing buffer (#781)
## Summary
This PR renames the `len` field in `InternerStorage` to `offset` to
better indicate the actual purpose of the field and adds a new field
called `len` which contains the size of all active entries.

## Change Type
- [x] Bug fix
- [ ] New feature
- [ ] Non-functional (chore, refactoring, docs)
- [ ] Performance


## How did you test this PR?
Added a unit test to check that `len` reflects the number of bytes
occupied by active entries.



## References
Closes: Track length of interner separate from offset into backing
buffer. <https://github.com/DataDog/saluki/issues/774>
<!-- Please list any issues closed by this PR. -->

<!--
- Closes: <issue link>
-->

<!-- Any other issues or PRs relevant to this PR? Feel free to list them
here. -->

---------

Co-authored-by: Raymond Zhao <35050708+rayz@users.noreply.github.com>