chore: update `RefreshableConfiguration` to use tls config with certificates (#809)
## Summary
<!-- Please provide a brief summary about what this PR does.
This should help the reviewers give feedback faster and with higher
quality. -->
We are seeing error logs, `"Failed to retrieve configuration..."` since
we need to use the `ipc_cert` now. This pr updates
`RefreshableConfiguration` to do so.

I moved the helper functions in `tonic.rs` to a new file in `saluki-io`
`ipc_.rs`.

## Change Type
- [ ] Bug fix
- [ ] New feature
- [x] Non-functional (chore, refactoring, docs)
- [ ] Performance

## How did you test this PR?
<!-- Please how you tested these changes here -->
Run ADP with the agent, the failure logs should be gone.
## References

<!-- Please list any issues closed by this PR. -->

<!--
- Closes: <issue link>
-->
- Closes: #808
<!-- Any other issues or PRs relevant to this PR? Feel free to list them
here. --> 4388102a71d455f7c187380c53970798b36ab8c4