enhancement: add origin detection support to DogStatsD events and service checks (#804)
## Summary
<!-- Please provide a brief summary about what this PR does.
This should help the reviewers give feedback faster and with higher
quality. -->


This PR introduces a new `TagsResolver` which contains the
`OriginTagsResolver`. This `TagsResolver` lives inside `ContextResolver`
and is created if not supplied to `ContextResolverBuilder`.

`Metrics`, `Events` and `ServiceChecks` use this `TagsResolver` in order
to get `OriginTags` as well as to create `TagSets`.

Further refactoring work is needed since `ContextResolver` still has
some logic involving `TagSetKey`.


## Change Type
- [ ] Bug fix
- [x] New feature
- [ ] Non-functional (chore, refactoring, docs)
- [ ] Performance


## How did you test this PR?
<!-- Please how you tested these changes here -->

## References

<!-- Please list any issues closed by this PR. -->

<!--
- Closes: <issue link>
-->

<!-- Any other issues or PRs relevant to this PR? Feel free to list them
here. -->