enhancement(core): add support for encoders and forwarders to topology blueprint (#811)
## Summary

This PR wires up encoders and forwarders into the topology machinery such that they can be added to `TopologyBlueprint` and spawned. In order to clean up a lot of the existing code that did not align with encoders/forwarders, we've done a lot of refactoring:

- created a new `TopologyContext` type (included in `SourceContext`, `TransformContext`, etc) that holds references to generic topology-level helpers, like the global thread pool and memory limiter
- added support to `TopologyBlueprint` to add encoders and forwarders (duh!)
- refactored much of the spawning code in `BuiltTopology` to more concisely build component interconnects, as well as spawn components

## Change Type
- [ ] Bug fix
- [x] New feature
- [ ] Non-functional (chore, refactoring, docs)
- [ ] Performance

## How did you test this PR?

Existing unit tests, and built ADP locally to ensure the current primary/internal observability topologies can be built and spawned, and that they function correctly.

## References

AGTMETRICS-233 fa01ba4a6f2e007ee62c07338f696038c393cd9f