enhancement(core, datadog): refactor HTTP payload type to be amenable to encoders (#813)
## Summary

This PR includes some refactoring around `ChunkedBytesBuffer`/`FrozenChunkedBytesBuffer` to make it more amenable to being used in HTTP payloads (`Payload::Http`).

Firstly, we've moved the types from `saluki-io` to `saluki-common` so that they can be used within `saluki-core`. Secondly, we've ripped out all of the buffer pool logic used to acquire additional chunks when more capacity is required. This follows the same logic as event buffers, where we know we're limited overall by the topology and have previously gotten bit by the buffer pools being undersized relative to how destinations were configured. This was a good time to make `ChunkedBytesBuffer` work similarly and so that's what we've done here.

Almost all of the other changes are related to removing buffer pool logic from `RequestBuilder`, updating the Datadog destinations, and so on.

## Change Type

- [ ] Bug fix
- [ ] New feature
- [x] Non-functional (chore, refactoring, docs)
- [ ] Performance

## How did you test this PR?

Existing unit tests.

## References

AGTMETRICS-233