enhancement(datadog): add Datadog forwarder (#815)
## Summary

This PR adds a basic Datadog forwarder component, based entirely on `TransactionForwarder`. We do the minimal amount of translation from `Payload` to `Transaction` and take advantage of the existing forwarder code used by destination.

In a follow-on PR, we'll actually wire up the DD Events encoder and the Datadog forwarder in the topology to replace the DD Events destination.

## Change Type

- [ ] Bug fix
- [x] New feature
- [ ] Non-functional (chore, refactoring, docs)
- [ ] Performance

## How did you test this PR?

N/A

## References

AGTMETRICS-233 a0a963a0bb1003f205b301f2afe9775cf1ab9c09