enhancement(agent-data-plane): switch to using Datadog Events encoder + Datadog forwarder (#816)
## Summary

This PR switches over to actually using the new Datadog Events encoder and Datadog forwarder. For now, only the Datadog Events encoder is flowing into the forwarder, but more things will be ported over to it as we convert existing destinations to encoders.

## Change Type

- [ ] Bug fix
- [ ] New feature
- [x] Non-functional (chore, refactoring, docs)
- [ ] Performance

## How did you test this PR?

Built and ran ADP locally and sent a number of test events to the DSD source. Observed the new encoder/forwarder components process and handle the event/payload, ultimately shipping it to the configured endpoint.

## References

AGTMETRICS-233 09f97a3fa193fc5495b225e354efcd177a387085