fix(dogstatsd): set hostname via 'host' well-known tag if available (#864)
## Summary

This PR updates the DogStatsD source to use the new "well-known tags" support to set the hostname for DSD metrics, as well as for calculating the origin of metrics, events, and service checks.

Prior to this PR, setting the `host` tag on a DSD metric did not cause the hostname of that metric to be properly overridden, as we were not extracting the `host` tag at all. We're now doing just that. This means we properly override the hostname we would otherwise set via the Host Enrichment transform, and we _also_ support disabling sending a hostname with metrics by setting an empty `host` tag (`host:`) on DSD metrics, allowing for cardinality control.

We've also wired up the well-known tags to help derive origin information for metrics, events, and service checks. This means we're utilizing WKT for both entity ID (pod UID) and cardinality (`card:`-based value is preferred if available). We'll remove the direct extraction of those tags/values in the DogStatsD codec as part of a follow-up PR.

## Change Type

- [x] Bug fix
- [ ] New feature
- [ ] Non-functional (chore, refactoring, docs)
- [ ] Performance

## How did you test this PR?

Sent counters/distributions to both the Datadog Agent and ADP, with the following variations: no host tag, an empty `host` tag (`host:`), and an overridden host tag. After this PR, the metrics sent through ADP matched those sent through the Datadog Agent in terms of how they arrived at and showed up as in the Datadog platform.

## References

AGTMETRICS-233