Add broadcast for config updates (#882)
## Summary
<!-- Please provide a brief summary about what this PR does.
This should help the reviewers give feedback faster and with higher
quality. -->

POC for adding a similar mechanism to the `OnUpdate` used in the
core-agent. It is not quite the same since we are not using callbacks
here.

This pr updates `run.rs` so that a bootstrap configuration is created.
If dynamic configuration is enabled, a `mpsc::channel` is created. The
sender and bootstrap config is used for
`create_config_stream(bootstrap_config, sender)`.

The `receiver` is used in the background task
`run_dynamic_config_updater` which is spawned when `into_generic` is
called.

This task is responsible for:
- receiving the initial snapshot and using `Notify` to signal that it
has been received. (A new method `ready` has been added to
`GenericConfiguration` and it will block until the signal has been
received. This is only used if dynamic configuration is enabled.)
- listening for all subsequent updates on a mpsc channel, diffing the
changes against the current state, and broadcasting `ConfigChangeEvents`
to any subscribers.

Configuration priority is determined by the order of which the providers
are added.

Component OnUpdate Mechanism: This system enables components to react to
changes. By calling `config.subscribe_for_updates()`, a component can
receive a stream of `ConfigChangeEvents`. It can then listen for changes
to relevant keys and update its internal state accordingly.
 


Example: https://github.com/DataDog/saluki/pull/883 (edit: outdated)



## Change Type
- [ ] Bug fix
- [x] New feature
- [ ] Non-functional (chore, refactoring, docs)
- [ ] Performance


## How did you test this PR?
<!-- Please how you tested these changes here -->
1. Run ADP with the agent
2. Run `config` subcommand in ADP
3. Change some configs
4. Run `config` subcommand in ADP
## References

<!-- Please list any issues closed by this PR. -->

<!--
- Closes: <issue link>
-->

<!-- Any other issues or PRs relevant to this PR? Feel free to list them
here. -->