[AGTMETRICS-310] add support for streaming configuration updates from the datadog agent (#797)
## Summary
<!-- Please provide a brief summary about what this PR does.
This should help the reviewers give feedback faster and with higher
quality. -->

This PR adds the implementation for streaming configuration events from
the core agent. This is gated behind the
`DD_ADP_USE_NEW_CONFIG_STREAM_ENDPOINT` environment variable. If
present, ADP will block until a snapshot has been retrieved via the
`ready` in `GenericConfiguration`.

Implementation:

`RemoteAgentClient` is used to establish a stream with the core agent
and a background task handles sending the updates to the
`GenericConfiguration` which has its own task
`run_dynamic_config_updater` (which is only run when
`with_dynamic_configuration` option is used. This task is responsible
for maintaining a copy of the dynamic state, and the current state so
that a diff can be produced. Each change is then sent to any subscribers
of `broadcast::Sender<ConfigChangeEvent>`. Note: This pr doesn't add any
subscribers yet, it only provides the functionality.

Providers are merged in based on the order they were added. For our
setup:
```
ConfigurationLoader::default()
            .try_from_yaml(&run_config.config)
            .with_dynamic_configuration(receiver)
            .from_environment("DD")?
            .with_default_secrets_resolution()
 ```
 Environment variables will have the highest priority over values retrieved dynamically here.



A new `config` subcommand has been added, along with a scrubber that hides sensitive values.


## Notes
1. The old method of retrieving up-to-date configuration settings via `RefreshableConfiguration` has been removed.
2. This branch served as the base for setting up the configuration stream. Relevant prs that have been since merged into this one:
   https://github.com/DataDog/saluki/pull/882
   https://github.com/DataDog/saluki/pull/842
   https://github.com/DataDog/saluki/pull/810

## Change Type
- [ ] Bug fix
- [x] New feature
- [ ] Non-functional (chore, refactoring, docs)
- [ ] Performance


## How did you test this PR?
<!-- Please how you tested these changes here -->

1. Manual test, updating configs and running the new `config` subcommand.
2. Unit tests for config diffs

## References

<!-- Please list any issues closed by this PR. -->

<!--
- Closes: <issue link>
-->

<!-- Any other issues or PRs relevant to this PR? Feel free to list them here. -->

---------

Co-authored-by: brianna-wang-dd <brianna.wang@datadoghq.com>