chore(saluki-config): add `ConfigurationLoader::for_tests` (#908)
## Summary
<!-- Please provide a brief summary about what this PR does.
This should help the reviewers give feedback faster and with higher
quality. -->
This pr adds a method `ConfigurationLoader::for_tests` and also a bunch
of unit tests that uses this method.

The dynamic updater has also been updated to handle nested keys that use
"`.`". Previously, it would just create the key on the top level of the
map object.

## Change Type
- [ ] Bug fix
- [ ] New feature
- [x] Non-functional (chore, refactoring, docs)
- [ ] Performance


## How did you test this PR?
<!-- Please how you tested these changes here -->
Unit tests
## References

<!-- Please list any issues closed by this PR. -->

<!--
- Closes: <issue link>
-->

<!-- Any other issues or PRs relevant to this PR? Feel free to list them
here. -->