chore(saluki-env): loop config stream connection (#926)
## Summary
<!-- Please provide a brief summary about what this PR does.
This should help the reviewers give feedback faster and with higher
quality. -->
This pr makes it so that we loop when trying to establish the config
stream connection to the core agent. This is so that a new connection is
created if the sender is dropped for whatever reason.

Also, this pr adds a change to use the default values for
`auth_token_file_path` and `ipc_cert_file_path` when they are `""`. This
is because the core-agent sets them as `""` so technically the value
exists, but not really.

## Change Type
- [x] Bug fix
- [ ] New feature
- [ ] Non-functional (chore, refactoring, docs)
- [ ] Performance


## How did you test this PR?
<!-- Please how you tested these changes here -->
1. Created agent custom image with the changes
[here](https://github.com/DataDog/datadog-agent/pull/41279) and deployed
it on an experimental cluster (sasquatch) with ADP enabled.
2. Run `agent-data-plane config` and see the config.
3. Update a setting (ex: `agent config set dogstatd_stats true`)
4. . Run `agent-data-plane config` and see the updated config value.
## References

<!-- Please list any issues closed by this PR. -->

<!--
- Closes: <issue link>
-->

<!-- Any other issues or PRs relevant to this PR? Feel free to list them
here. -->