chore(agent-data-plane): add configuration interlock to exit if ADP not enabled (#938)
## Summary

This PR adds a small bit of logic around ADP's startup in order to exit if ADP is not enabled.

As we push towards bundling ADP with the Datadog Agent, we need a more direct mechanism to allow ADP to startup and determine if it should be running, such that this configuration is able to flow from the Core Agent (eventually "control plane") directly.

We now evaluate a new configuration setting -- `data_plane.enabled` -- which, when `false`, causes ADP to immediately exit. This setting defaults to `false`. We add an escape hatch to ignore it when running in standalone mode, to handle local development and benchmarks.

The flow will be that when ADP should be enabled, `data_plane.enabled`/`DD_DATA_PLANE_ENABLED` will need to be set to `true`. This is compatible with how we currently deploy ADP (insofar as we can update our deployment flow to set this new setting _today_) and compatible with the non-standalone-mode approach of sourcing our configuration from the Core Agent/control plane.

## Change Type

- [ ] Bug fix
- [x] New feature
- [ ] Non-functional (chore, refactoring, docs)
- [ ] Performance

## How did you test this PR?

Built and ran ADP locally, in six different ways:

- non-standalone mode, `DD_DATA_PLANE_ENABLED` not set
- non-standalone mode, `DD_DATA_PLANE_ENABLED` set to `false`
- non-standalone mode, `DD_DATA_PLANE_ENABLED` set to `true`
- standalone mode, `DD_DATA_PLANE_ENABLED` not set
- standalone mode, `DD_DATA_PLANE_ENABLED` set to `false`
- standalone mode, `DD_DATA_PLANE_ENABLED` set to `true`

In non-standalone mode when the setting was not set, or was set explicitly to `false`, ADP exited as intended. In the other four cases (setting set to `true`, or standalone mode enabled), ADP started up as intended.

## References

AGTMETRICS-233