refactor(app): start splitting logging config into standalone struct (#948)
## Summary

This PR is the first of many related to a newer, all-in-one "bootstrap" procedure for initializing data plane applications. This proposed bootstrap procedure is meant to act as a single entrypoint for all core feature initialization, including unified reading of settings from configuration, and so on.

As part of preparing for that, we're going through and porting all existing configuration of current initialization functions into dedicated configuration structs, such that we can load their configuration in a unified way through `GenericConfiguration`.

This PR introduces `LoggingConfiguration` to be used for logging initialization. Practically, we're using the struct under the hood by creating a standalone `GenericConfiguration` after reading environment variables, and then deserializing the configuration. As we start exposing a "unified" bootstrap procedure, we'll move to expose these types and remove the standalone `initialize_logging` function.

Additionally, we've remove the usage of `initialize_logging` from non-ADP binaries. In order to create `GenericConfiguration`, we have to call an async function, which means our "under the hood" approach requires `async`. We've moved `initialize_dynamic_logging` (which is already `async`) to be `initialize_logging`, collapsing down to a single initialization path for logging. Since we don't want to force that on the correctness binaries, and they don't actually _need_ identical logging... we just switched to using `tracing_subscriber` directly for those.

## Change Type

- [ ] Bug fix
- [ ] New feature
- [x] Non-functional (chore, refactoring, docs)
- [ ] Performance

## How did you test this PR?

Built and ran ADP and the various correctness testing binaries by hand. Ensured that existing logging configurability for ADP works as expected, and that logging _generally_ worked for the correctness testing binaries.

## References

AGTMETRICS-233 e23aed831f917501aa1e1d24a18498ab28c32983