enhancement(app): add support to log to file (#949)
## Summary

This PR adds support for writing logs to a log file. We follow the existing configuration setting names used in the Datadog Agent for controlling log file-based logging:

- `log_file`: path to log file (logging to file disabled when path is empty)
- `log_file_max_size`: size in bytes before rotating (either raw bytes or human-friendly like `1 KiB` or `5Mb`; defaults to `10Mb`)
- `log_file_max_rolls`: maximum number of rolled-over files in addition to main log file (defaults to `1`)

The vast majority of this PR is rearranging the existing logger setup code. Since we now have multiple "backends", we want to share some of the common stuff (like JSON vs pretty printed) as well as switching to boxed layers to avoid needing complex enums for concrete typing and all of that... which is a long way of saying that I moved a bunch of the formatting/layer-specific code into its own module, and then also moved the API-related bits to an API-specific module, similar to what we do in `saluki_app::metrics`.

## Change Type

- [ ] Bug fix
- [x] New feature
- [ ] Non-functional (chore, refactoring, docs)
- [ ] Performance

## How did you test this PR?

Built and ran ADP locally, setting `DD_LOG_FILE`, `DD_LOG_FILE_MAX_SIZE`, and `DD_LOG_FILE_MAX_ROLLS` to ensure each aspect worked correctly.

## References

AGTMETRICS-233